chore(developer): move additional consts and globals
